renamed and relocated auth guard, created role guard, created bulk-upload routing reference
