+++ src/app/app.module.ts
@@ -10,7 +10,7 @@ import {
 import { ServiceUnavailableComponent } from '@core/components/error/service-unavailable/service-unavailable.component';
 import { FooterComponent } from '@core/components/footer/footer.component';
 import { HeaderComponent } from '@core/components/header/header.component';
-import { AuthGuard } from '@core/services/auth-guard.service';
+import { AuthGuard } from '@core/guards/auth/auth.guard';
 import { AuthInterceptor } from '@core/services/auth-interceptor';
 import { BackService } from '@core/services/back.service';
 import { CountryService } from '@core/services/country.service';
