+++ src/app/core/guards/auth/auth.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, CanActivateChild, Router, RouterStateSnapshot } from '@angular/router';
 
-import { AuthService } from './auth.service';
+import { AuthService } from '@core/services/auth.service';
 
 @Injectable({
   providedIn: 'root',
