+++ src/app/core/guards/role/role.guard.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, async, inject } from '@angular/core/testing';
+
+import { RoleGuard } from './role.guard';
+
+describe('RoleGuard', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [RoleGuard]
+    });
+  });
+
+  it('should ...', inject([RoleGuard], (guard: RoleGuard) => {
+    expect(guard).toBeTruthy();
+  }));
+});
